backend/requirements.txt
appnope==0.1.3
asgiref==3.5.0
asttokens==2.0.8
backcall==0.2.0
debugpy==1.6.3
decorator==5.1.1
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
entrypoints==0.4
executing==1.0.0
ipykernel==6.15.3
ipython==8.5.0
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.3.5
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
psutil==5.9.2
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
pyzmq==24.0.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.5.0
tornado==6.2
traitlets==5.4.0
wcwidth==0.2.5
whitenoise==6.0.0
